+++ Sources/Controllers/ALKConversationViewController.swift
@@ -1020,16 +1020,6 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         navigationController?.pushViewController(conversationVC, animated: true)
     }
 
-    func menuItemSelected(action: ALKChatBaseCell<ALKMessageViewModel>.MenuActionType,
-                          message: ALKMessageViewModel) {
-        switch action {
-        case .reply:
-            print("Reply selected")
-            viewModel.setSelectedMessageToReply(message)
-            replyMessageView.update(message: message)
-            showReplyMessageView()
-        }
-    }
 
     func showReplyMessageView() {
         replyMessageView.constraint(
@@ -1911,4 +1901,71 @@ extension ALKConversationViewController: NavigationBarCallbacks {
         return ALContactService().loadContact(byKey: "userId", value: contactId)
     }
 
+}
+
+extension ALKConversationViewController: ALAlertButtonClickProtocol {
+
+    func confirmButtonClick(action: String,messageKey: String) {
+        let alPushAssist =  ALPushAssist()
+
+        if(action == ALKAlertViewController.Action.reportMessage){
+
+            alPushAssist.topViewController.dismiss(animated: false, completion: nil)
+
+            guard  ALDataNetworkConnection.checkDataNetworkAvailable() else {
+                return
+            }
+
+            let userService = ALUserService()
+            let activityIndicator = UIActivityIndicatorView(style: UIActivityIndicatorView.Style.gray)
+            activityIndicator.center = CGPoint(x: view.bounds.size.width/2,
+                                               y: view.bounds.size.height/2)
+            activityIndicator.color = UIColor.gray
+            view.addSubview(activityIndicator)
+            activityIndicator.startAnimating()
+
+            let message = localizedString(forKey: "ReportMessageSuccess", withDefaultValue: SystemMessage.Information.ReportMessageSuccess, fileName: configuration.localizedStringFileName)
+
+            let errorMessage = localizedString(forKey: "ReportMessageError", withDefaultValue: SystemMessage.Information.ReportMessageError, fileName: configuration.localizedStringFileName)
+
+            userService.reportUser(withMessageKey: messageKey) { (apiResponse, error) in
+                activityIndicator.stopAnimating()
+                if(error == nil){
+                    self.showAlert(alertTitle: "", alertMessage: message)
+                }else{
+                    self.showAlert(alertTitle: "", alertMessage: errorMessage)
+                }
+            }
+        }
+    }
+
+    func showAlert(alertTitle:String,alertMessage:String)  {
+        let alPushAssist =  ALPushAssist()
+        let title = localizedString(forKey: "OkMessage", withDefaultValue: SystemMessage.ButtonName.ok, fileName: configuration.localizedStringFileName)
+        let alert = UIAlertController(
+            title: alertTitle,
+            message: alertMessage,
+            preferredStyle: UIAlertController.Style.alert)
+        alert.addAction(UIAlertAction(title: title, style: UIAlertAction.Style.default, handler:nil))
+        alPushAssist.topViewController.present(alert, animated: true, completion: nil)
+    }
+
+
+    func menuItemSelected(action: ALKChatBaseCell<ALKMessageViewModel>.MenuActionType,
+                          message: ALKMessageViewModel) {
+        switch action {
+        case .reply:
+            print("Reply selected")
+            viewModel.setSelectedMessageToReply(message)
+            replyMessageView.update(message: message)
+            showReplyMessageView()
+        case .reportMessage:
+            let muteConversationVC = ALKAlertViewController(action:ALKAlertViewController.Action.reportMessage, delegate: self, messageKey: message.identifier,  configuration: configuration)
+            let title = localizedString(forKey: "ReportAlertTitle", withDefaultValue: SystemMessage.Information.ReportAlertTitle, fileName: configuration.localizedStringFileName)
+            let message = localizedString(forKey: "ReportAlertMessage", withDefaultValue: SystemMessage.Information.ReportAlertMessage, fileName: configuration.localizedStringFileName)
+            muteConversationVC.updateTitleAndMessage(title, message: message)
+            muteConversationVC.modalPresentationStyle = .overCurrentContext
+            self.present(muteConversationVC, animated: true, completion: nil)
+        }
+    }
 }