+++ Sources/Utilities/SystemMessage.swift
@@ -46,6 +46,10 @@ struct SystemMessage : Localizable {
         static let UnblockToEnableChat = localizedString(forKey: "UnblockToEnableChat")
         static let ChatHere = localizedString(forKey: "ChatHere")
         static let LoadingIndicatorText = localizedString(forKey: "LoadingIndicatorText")
+        static let ReportMessageSuccess = localizedString(forKey: "ReportMessageSuccess")
+        static let ReportMessageError = localizedString(forKey: "ReportMessageError")
+        static let ReportAlertTitle = localizedString(forKey: "ReportAlertTitle")
+        static let ReportAlertMessage = localizedString(forKey: "ReportAlertMessage")
     }
 
     struct Update {
@@ -64,6 +68,7 @@ struct SystemMessage : Localizable {
         static let CamNotAvaiable = localizedString(forKey: "CamNotAvaiable")
         static let Cancelled = localizedString(forKey: "CancelMessage")
         static let PleaseTryAgain = localizedString(forKey: "ConnectionFailedMessage")
+        static let NetworkAccessFailedMessage = localizedString(forKey: "NetworkAccessFailedMessage")
         static let FetchFail = localizedString(forKey: "FetchFailedMessage")
         static let OperationFail = localizedString(forKey: "OperationFailedMessage")
         static let DeleteSingleConversation = localizedString(forKey: "DeleteSingleConversation")
@@ -86,6 +91,7 @@ struct SystemMessage : Localizable {
         static let Remove = localizedString(forKey: "RemoveButtonName")
         static let Leave = localizedString(forKey: "LeaveButtonName")
         static let Cancel = localizedString(forKey: "ButtonCancel")
+        static let CapitalLetterCancelText = localizedString(forKey: "CapitalLetterCancelText")
         static let Discard = localizedString(forKey: "ButtonDiscard")
         static let Save = localizedString(forKey: "SaveButtonTitle")
         static let ResetPhoto = localizedString(forKey: "ResetPhotoButton")
@@ -93,6 +99,7 @@ struct SystemMessage : Localizable {
         static let Done = localizedString(forKey: "DoneButton")
         static let Invite = localizedString(forKey: "InviteButton")
         static let Confirm = localizedString(forKey: "ConfirmButton")
+        static let ReportMessage = localizedString(forKey: "ReportMessage")
         static let Delete = localizedString(forKey: "DeleteButtonName")
         static let ok = localizedString(forKey: "OkMessage")
     }
@@ -153,6 +160,7 @@ struct SystemMessage : Localizable {
         static let NotNow = localizedString(forKey: "NotNow")
         static let Copy = localizedString(forKey: "Copy")
         static let Reply = localizedString(forKey: "Reply")
+        static let Report = localizedString(forKey: "Report")
         static let You = localizedString(forKey: "You")
         static let Admin = localizedString(forKey: "Admin")
     }