+++ Sources/Views/ALKBaseCell.swift
@@ -12,7 +12,6 @@ import UIKit
 open class ALKBaseCell<T>: UITableViewCell {
 
     var viewModel: T?
-
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         setupViews()